#111 Server - Domain - Use of reflection to check for an empty struct and not found errors

+++ repertoire.server/domain/usecase/playlist/remove_song_from_playlist.go
@@ -1,6 +1,8 @@
 package playlist
 
 import (
+	"errors"
+	"reflect"
 	"repertoire/server/data/repository"
 	"repertoire/server/internal/wrapper"
 	"repertoire/server/model"
@@ -29,11 +31,17 @@ func (r RemoveSongFromPlaylist) Handle(id uuid.UUID, songID uuid.UUID) *wrapper.
 	if err != nil {
 		return wrapper.InternalServerError(err)
 	}
+	if reflect.ValueOf(playlist).IsZero() {
+		return wrapper.NotFoundError(errors.New("playlist not found"))
+	}
 
 	var song model.Song
 	err = r.songRepository.Get(&song, songID)
 	if err != nil {
 		return wrapper.InternalServerError(err)
+	}
+	if reflect.ValueOf(song).IsZero() {
+		return wrapper.NotFoundError(errors.New("song not found"))
 	}
 
 	err = r.repository.RemoveSong(&playlist, &song)
